+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 1.0.0-beta.15
 ### Features
 * improve performance of ** operator when JS support exponential operator
+### Bugfix
+* fix scoping issue in nested `syntax-rules`
 
 ## 1.0.0-beta.14
 ### Breaking